requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Babel==4.0.0
Flask-WTF==1.2.1
Authlib==1.3.0
python-dotenv==1.0.0
requests==2.31.0

# 测试依赖
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# 性能监控
psutil==5.9.0

